mais do backend

+++ Backend-ecommerce/src/main/java/com/example/ecommerce/repository/CategoriaRepository.java
@@ -3,5 +3,5 @@ package com.example.ecommerce.repository;
 import com.example.ecommerce.entity.Categorias;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoriaRepository extends JpaRepository<Categorias, Long> {
+public interface CategoriaRepository extends JpaRepository<Categorias, Integer> {
 }
